pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sregym"
version = "0.1.0"
description = "SRE incident-triage RL environment"
requires-python = ">=3.11"
dependencies = [
    "fastapi~=0.115.0",
    "uvicorn~=0.30.0",
]

[project.optional-dependencies]
coordinator = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25",
]
dev = [
    "sregym[coordinator]",
    "pytest>=7.0",
]

[project.scripts]
sregym = "sregym.cli:main"

[tool.setuptools.packages.find]
include = ["sregym*"]

[tool.pyright]
include = ["sregym", "tests"]
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
